requirements.txt
# extracted from a notebook

## third-party
crunch-cli  # alias of crunch
endersgame
matplotlib
numpy~=2.0.2
pandas~=2.2.3
scipy~=1.13.1
tqdm~=4.66.5

## standard
#datetime
#json
#os
#typing

pydantic~=2.9.2
nevergrad~=1.0.5
pytest~=6.2.5
midplot~=0.1.0
midone~=1.0.3
ipywidgets~=8.1.5
ipython~=8.18.1
plotly~=5.24.1
setuptools~=75.2.0
